spring

+++ sts/Spring-Jwt/src/main/java/com/example/demo/repository/usersRepository.java
@@ -3,7 +3,7 @@ package com.example.demo.repository;
 import com.example.demo.model.users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface usersRepository extends JpaRepository<users,Long>{
+public interface usersRepository extends JpaRepository<users,Integer>{
 	//username, it will return the user of the given username
 	public users findByUsername(String username);
 }
